Removed SoftwareSerial and TMC2208Stepper.cpp
In TMCStepper.h:
SoftwareSerial has been causing some bugs when I use this library with my plottrbot. All features of SoftwareSerial has been removed/ignored in my fork of the TMCStepper library.

TMC2208Stepper.cpp has been renamed to be ignored. This file had some depencies on softwareserial, it was easiest to just ignore this as I'm not using the TMC2208 drivers.